Moved commonly used test-setup method to utility class
