+++ oscm-rest-api-account/src/test/java/org/oscm/rest/account/BillingContactResourceTest.java
@@ -20,11 +20,10 @@ import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.oscm.rest.account.data.BillingContactRepresentation;
 import org.oscm.rest.common.RepresentationCollection;
+import org.oscm.rest.common.SampleTestDataUtility;
 
-import javax.ws.rs.core.*;
-import java.net.URI;
-import java.util.Collections;
-import java.util.List;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -45,7 +44,7 @@ public class BillingContactResourceTest {
   @BeforeEach
   public void setUp() {
     representation = createBillingContactRepresentation();
-    uriInfo = createUriInfo();
+    uriInfo = SampleTestDataUtility.createUriInfo();
     parameters = createParameters();
   }
 
@@ -152,109 +151,6 @@ public class BillingContactResourceTest {
     BillingContactRepresentation representation = new BillingContactRepresentation();
     representation.setId(123L);
     return representation;
-  }
-
-  private UriInfo createUriInfo() {
-    return new UriInfo() {
-      @Override
-      public String getPath() {
-        return null;
-      }
-
-      @Override
-      public String getPath(boolean b) {
-        return null;
-      }
-
-      @Override
-      public List<PathSegment> getPathSegments() {
-        return null;
-      }
-
-      @Override
-      public List<PathSegment> getPathSegments(boolean b) {
-        return null;
-      }
-
-      @Override
-      public URI getRequestUri() {
-        return null;
-      }
-
-      @Override
-      public UriBuilder getRequestUriBuilder() {
-        return null;
-      }
-
-      @Override
-      public URI getAbsolutePath() {
-        return null;
-      }
-
-      @Override
-      public UriBuilder getAbsolutePathBuilder() {
-        return null;
-      }
-
-      @Override
-      public URI getBaseUri() {
-        return null;
-      }
-
-      @Override
-      public UriBuilder getBaseUriBuilder() {
-        return null;
-      }
-
-      @Override
-      public MultivaluedMap<String, String> getPathParameters() {
-        return new MultivaluedHashMap<String, String>() {
-          {
-            put("version", Collections.singletonList("v1"));
-          }
-        };
-      }
-
-      @Override
-      public MultivaluedMap<String, String> getPathParameters(boolean b) {
-        return null;
-      }
-
-      @Override
-      public MultivaluedMap<String, String> getQueryParameters() {
-        return null;
-      }
-
-      @Override
-      public MultivaluedMap<String, String> getQueryParameters(boolean b) {
-        return null;
-      }
-
-      @Override
-      public List<String> getMatchedURIs() {
-        return null;
-      }
-
-      @Override
-      public List<String> getMatchedURIs(boolean b) {
-        return null;
-      }
-
-      @Override
-      public List<Object> getMatchedResources() {
-        return null;
-      }
-
-      @Override
-      public URI resolve(URI uri) {
-        return null;
-      }
-
-      @Override
-      public URI relativize(URI uri) {
-        return null;
-      }
-    };
   }
 
   private AccountParameters createParameters() {
